+++ commands/rollm.js
@@ -1,13 +1,13 @@
 module.exports = {
-	name: 'rolla',
-	description: 'Rolls dice for the *ALIEN* roleplaying game.'
+	name: 'rollm',
+	description: 'Rolls dice for the *Mutant: Year Zero* roleplaying game.'
 		+ 'Type `help roll` for more details.',
-	aliases: ['ra', 'lancea', 'lancera', 'slåa', 'slaa'],
+	aliases: ['rm', 'lancem', 'lancerm', 'slåm', 'slam'],
 	guildOnly: false,
 	args: true,
 	usage: '<dice> [arguments]',
 	execute(args, message, client) {
-		args.unshift('alien');
+		args.unshift('myz');
 		client.commands.get('roll').execute(args, message, client);
 	},
 };